Split waste_policies.rs (820 lines) into subdirectory modules
Break the monolithic waste_policies.rs into smaller, focused files:
- constants.rs (38 lines): policy cost/bonus/penalty constants
- state.rs (194 lines): WastePolicyState, WastePolicyEffects, helper functions, Saveable impl
- systems.rs (111 lines): update_waste_policies system and WastePoliciesPlugin
- tests_state.rs (341 lines): state defaults, helpers, costs, and saveable tests
- tests_effects.rs (121 lines): WTE diversion and combined effects tests
- mod.rs (30 lines): module docs and re-exports

All public items are re-exported from mod.rs so external callers are unaffected.

Part of #1389
